requirements.txt
gradio==4.44.1
gradio-client==1.3.0
fastapi==0.112.2
starlette==0.38.6
pydantic==2.8.2
huggingface_hub==0.24.7
sympy==1.13.1

torch
torchvision
timm
numpy
pillow
opencv-python
pandas
matplotlib
